🚧 (wip) extract chart state from discrete bar charts
